Error Handling
Fix error handling for tasks with index

There are several methods in TaskList which require user input for line
that are not sync. Error message are not updated appropriately.

Fixing errors would help debug smoothly, but more importantly ensure
code works as intended.

Let's change the condition for methods in Tasklist that reference index,
and throw the correct error with an appropriate message for
getResponse() in MainWindow to refer to, and display the error message
to the user.